refactor(gateway): extract session-resume authorization into GatewayResumeAuthorizationMixin (god-file Phase 3)
GatewaySlashCommandsMixin was itself lifted out of gateway/run.py by the
god-file decomposition campaign. It has since grown to 5,383 lines and 65
methods, which makes it a decomposition target in its own right.

This takes the cluster in it with the clearest boundary: deciding whether a
caller may resume, or even see, somebody else's session. That is a trust
decision rather than command plumbing. _resume_target_allowed is what stops a
user in one chat from resuming a session belonging to a user in another, and
tests/gateway/test_resume_command.py exercises it with explicit cross-user and
IDOR cases.

Moved verbatim (279 lines, 6 methods, contiguous at 912-1195):

  _gateway_session_origin_for_id    _resume_caller_is_admin
  _same_matrix_room                 _resume_target_allowed
  _same_origin_chat                 _resume_row_visible

Behavior-neutral. Every body is byte-identical to what it replaced; the only
edits are the class declaration, one import, and the new module header.

The new mixin becomes a base of GatewaySlashCommandsMixin rather than of
GatewayRunner, so gateway/run.py is untouched and the change stays inside one
file plus the new module. GatewayRunner picks the methods up through the
existing MRO: GatewayRunner -> GatewayAuthorizationMixin ->
GatewayKanbanWatchersMixin -> GatewaySlashCommandsMixin ->
GatewayResumeAuthorizationMixin. Verified at runtime that every method name
defined on the old class still resolves off GatewayRunner.

The cluster calls no method that stays behind, and needs only Optional,
Platform, SessionSource and is_shared_multi_user_session, taken from their
original modules. _redact_matrix_session_key stays put: its one caller is
_handle_context_command, which is not part of this cluster.

slash_commands.py 5,483 -> 5,199 lines; the class 5,383 -> 5,098 lines and
65 -> 59 methods.
